Added OPTS.CLONE
MATLAB treats objects as references. Changing a reference affects the original object.
To reduce problems caused by this, the OPTS.CLONE function will copy the options in an
options argument into a new reference.

ISSUES: if the options object contains objects of their own, both the original and the
cloned object will point to the same reference.